# backend_openapi_v3.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "NECO Photo Album API",
                 "description":  "FastAPI backend for NECO photo album generation",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/students/{student_id}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "students"
                                                                             ],
                                                                    "summary":  "Get Student",
                                                                    "operationId":  "get_student_api_v1_students__student_id__get",
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "student_id",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string",
                                                                                                          "format":  "uuid",
                                                                                                          "title":  "Student Id"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Successful Response",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/StudentRead"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              },
                                                                                      "422":  {
                                                                                                  "description":  "Validation Error",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                },
                                                        "put":  {
                                                                    "tags":  [
                                                                                 "students"
                                                                             ],
                                                                    "summary":  "Update Student",
                                                                    "operationId":  "update_student_api_v1_students__student_id__put",
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "student_id",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string",
                                                                                                          "format":  "uuid",
                                                                                                          "title":  "Student Id"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "requestBody":  {
                                                                                        "required":  true,
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/StudentUpdate"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Successful Response",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/StudentRead"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              },
                                                                                      "422":  {
                                                                                                  "description":  "Validation Error",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                },
                                                        "delete":  {
                                                                       "tags":  [
                                                                                    "students"
                                                                                ],
                                                                       "summary":  "Delete Student",
                                                                       "operationId":  "delete_student_api_v1_students__student_id__delete",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "student_id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "string",
                                                                                                             "format":  "uuid",
                                                                                                             "title":  "Student Id"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {

                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                    },
                  "/api/v1/students/by-state/{state_code}":  {
                                                                 "get":  {
                                                                             "tags":  [
                                                                                          "students"
                                                                                      ],
                                                                             "summary":  "Get Students By State",
                                                                             "operationId":  "get_students_by_state_api_v1_students_by_state__state_code__get",
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "state_code",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string",
                                                                                                                   "title":  "State Code"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "Successful Response",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "type":  "array",
                                                                                                                                                                   "items":  {
                                                                                                                                                                                 "$ref":  "#/components/schemas/StudentWithSchoolDetails"
                                                                                                                                                                             },
                                                                                                                                                                   "title":  "Response Get Students By State Api V1 Students By State  State Code  Get"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       },
                                                                                               "422":  {
                                                                                                           "description":  "Validation Error",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}